upgrade solve model, stat calculation to /stats endpoint

+++ backend/app/schemas/solve.py
@@ -4,20 +4,20 @@ from datetime import datetime
 class SolveSchema(BaseModel):
     id: int
     cube_type: str
-    time: int # stored in centiseconds
+    time: float
     scramble: str
     dnf: bool
     createdAt: datetime
 
 class SolveRequest(BaseModel):
     cube_type: str
-    time: int # stored in centiseconds
+    time: float
     scramble: str
     dnf: bool
 
 class StatsSchema(BaseModel):
-    best_ao5: int
-    best_ao12: int
-    best_ao100: int
-    best_time: int
+    best_ao5: float
+    best_ao12: float
+    best_ao100: float
+    best_time: float
     total_solves: int
